Issue #122: Issues while scheduling suites on parallel execution
--suitesfrom was using as a option to get suites to be run. This patch
propose a new approach to use it.

Create a option --orderfrom, which get the suites ordered from
output.xml, but, instead of only run them, merge the suites from passed
TAGS, which is good when you have a project with too much commits

1) get suites by tags
2) get suites by --orderfrom
3) append suites existent in tags, that are not existent in --suitesfrom
list
4) verify if new list has the same elements of suites by tags

So, this new way will use output.xml as a guide. Tags will be the main
list to be followed.

Signed-off-by: Ramon Medeiros <ramon.rnm@gmail.com>